Separate out model definitions
