+++ app/application.rb
@@ -4,19 +4,12 @@ require_relative 'tumblr_poller'
 require_relative 'tumblr_interface'
 require_relative 'overlap_service'
 
+require_relative './models/blog'
+require_relative './models/post'
+require_relative './models/comparison'
+
 # ActiveRecord::Base.logger = Logger.new(STDERR)
 ActiveRecord::Base.establish_connection(
   adapter: 'sqlite3',
   database: 'database.db'
 )
-
-class Blog < ActiveRecord::Base
-  has_many :posts
-end
-
-class Post < ActiveRecord::Base
-  belongs_to :blog
-end
-
-class Comparison < ActiveRecord::Base
-end
