+++ app/models/blog.rb
@@ -1,0 +1,3 @@
+class Blog < ActiveRecord::Base
+  has_many :posts
+end
