+++ app/models/comparison.rb
@@ -1,0 +1,4 @@
+require 'spec_helper'
+
+class Comparison < ActiveRecord::Base
+end
