+++ app/models/post.rb
@@ -1,0 +1,3 @@
+class Post < ActiveRecord::Base
+  belongs_to :blog
+end
