+++ spec/app/models/blog_spec.rb
@@ -1,0 +1,4 @@
+require 'spec_helper'
+
+RSpec.describe Blog do
+end
