+++ spec/app/models/comparison_spec.rb
@@ -1,0 +1,4 @@
+require 'spec_helper'
+
+RSpec.describe Comparison do
+end
