+++ spec/app/models/post_spec.rb
@@ -1,0 +1,4 @@
+require 'spec_helper'
+
+RSpec.describe Post do
+end
